Removing the docker login step that causes forked PRs to fail (#844)
Also adding pull_request_target event for release and master branch. This will run the GH Actions in the context of the base repository instead of the forked one. Hence, the secrets for Cypress, Docker etc can be passed down to the action for PRs being raised via a fork