fix: Show button when fetching columns (#6891)
Improve UX when entering `Table Header Index` value

BEFORE: Generate Template Button was hidden when the new column header is getting fetched. There was no loader for a user to understand that data is being fetched.

NOW: Generate Template Button is always visible and goes in a loading state when user data is being fetched.