feat: Add an API to get env values for superusers (#7043)
This commit adds the following API:

`/api/v1/admin/env` — Responds with the names and values of all environment variables currently configured, that are allowed for configuration from the UI. Some variables will be missing from this, ones that are not allowed to be edited from the configuration UI.

Doesn't take any query params. Only usable by the superuser.

Also added a `isSuperUser` boolean field to the response of `api/v1/users/me`, that will indicate whether the current user is a superuser or not.