Delay on page load actions till all actions are retrieved (#13657)
As a quick fix to prevent the on-page load actions from firing before the actions are even fetched, we had reverted the changes to make fetch page DSL call parallel in 13587. Unfortunately, theming PR brought back parallelisation and the issue along with it.

This PR ensures that the on-page-load actions are only made after actions are fetched while preserving parallelisation.