fix: fix debugger logs for performance (#16451)
This Change includes two parts,

In Debugger Saga, Action is being triggered in a loop, that action is in turn updating Redux State in a loop, This is bad for performance as they end up notifying the all the subscribers of the state. Changed it to update state once rather than doing it individually.
In the Reducer, changed it to ImmerReducer so that only particular object of the local state changes on update rather than updating reference of the entire state.