fix: forking template to application doesn't have page permissions (#19134)
## Description

When forking a template, we fetch the page list that comes with the
template. But we only fetch the first page, and not the rest. Hence we
won't have the permissions for the rest of the page. So the changes now
fetch the pages which we get from the template on fork.

Fixes #19127


## Type of change

> Please delete options that are not relevant.

- Bug fix (non-breaking change which fixes an issue)



## How Has This Been Tested?

- Manual


## Checklist:
### Dev activity
- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] New and existing unit tests pass locally with my changes
- [ ] PR is being merged under a feature flag