fix: invite user should not update existing member's role (#19254)
Fixes #19244 
Fixes #17606

If an existing member is invited again from the Share modal with a
different role, the user gets assigned the other role as well. This
breaks the framework that a member of the workspace can only have a
SINGLE role. This breaks other flows like update role/delete member etc
because they all run with the assumption that the member can not have
more than one role in the workspace. This PR fixes this by throwing an
error if a user is being invited again to the workspace.