fix: This fixes the DDOS that appsmith production database faced when the sessions set in redis were in bad state. (#19807)
1. Pre fetch and set the anonymous user cache instead of fetching it
when its required. This is now done on startup. In case any request for
anonymous user permission group ids comes through and the cache is not
ready, we throw an error to request the user to try again in some time.
This stops the DDOS on the mongo database
2. In case the session is in bad state and the user object is malformed,
before we fetch the user permission group ids, we check the presence of
email, tenant and user id before making a bad database query which was
another reason for DDOS.