chore:  Add photos of workspace members in homepage API response (#20132)
The homepage API is used to load list of workspaces, apps under those
workspaces and their members. The list of members is used to display a
profile photo of the members. The client loads the profile picture using
the email address of the member. This can lead to 404 if the member has
no photo set.

This PR includes the photo ids in the response so that client knows
whether the user has photo or not. If photo is not set, client will not
try to load it. So the 404 responses will no more be there.

Fixes #20