fix: Change watchtower to latest (#20001)
We're `watchtower:latest-dev` image for the `--schedule` support. This
was only available in the `latest-dev` version of Watchtower.

Now, this feature is available in the `latest` version of Watchtower, so
we should be fine switching back to `latest`.

More importantly, the `latest-dev` version of Watchtower, doesn't have
ARM images built. It only has AMD images built today. So this watchtower
service fails to start on ARM architectures. This PR fixes that.