chore: Adding environmentName parameter in datasource test flow and required decomposition and refactors. (#19783)
## Description
- TL; DR , small Pr to add environmentName to Datasource-controller for
`test`, `structure`, and `trigger` flows. In order to adapt for
environments. There is a method addition in authenticationValidator, and
further refactor to avoid code duplicity.

- With Introduction of multiple environments we would have
configurations corresponding to each environment. In order to test these
configurations, we would require an environment parameter to be passed
as an argument, this could be done by passing an environment header with
the current datasource `test`, `structure`, and `trigger` flows,
however, since our community and enterprise flow uses the same code. We
require our community code to include the header which however, will not
be used in the community edition but it will have complete usage in
enterprise edition.
- A `validateAuthentication` method in `authenticationValidator`
interface has been loaded with an extra parameter (environmentId) for
environment support, EE changes will follow post the merge.
- `Datasource trigger and Datasource Structure flows`, required some
refactors to make it work with environments as well.

## Type of change

- Chore (housekeeping or task changes that don't impact user perception)

## How Has This Been Tested?
- Manual
- Jest
- Cypress

Existing test cases are adequate, just tweaked a few test cases to
accommodate changes.


## Checklist:
### Dev activity
- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] New and existing unit tests pass locally with my changes
- [ ] PR is being merged under a feature flag


### QA activity:
- [ ] Test plan has been approved by relevant developers
- [ ] Test plan has been peer reviewed by QA
- [ ] Cypress test cases have been added and approved by either SDET or
manual QA
- [ ] Organized project review call with relevant stakeholders after
Round 1/2 of QA
- [ ] Added Test Plan Approved label after reveiwing all Cypress test